+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.2.3
 
+- [#440](https://github.com/okta/okta-auth-js/pull/440) Fixes signOut XHR fallback to reload page only if postLogoutRedirectUri matches the current URI
 - [#445](https://github.com/okta/okta-auth-js/pull/445) Clears access token from storage after token revocation
 
 ## 3.2.2